IndexFile Created and Tested
Created and Tested an IndexFile class which manages an index file for
the Cache, so instead of write 1000s of individual files to disk when
cache items are evicted, this class will manage an index file which can
be used to store the items key and the values position in the data file.
The data file wll be a single append only file which will hold all the
values.