Switched Direction
So the DiskCache didn't work out, it was very very very slow! So its
been deleted.

But I did abstract the FileCaching into an Interface ICacheToDisk and
tried another implementation which uses directories. The directory cache
is much faster than the File one. This is now the default implemetation.

Found a bug while actracting the DiskCache where the previous code was
checing for files in c:\ and not c:\baseDirectory, so it was constantly
saving items to the disk.